requirements.txt
alembic==1.16.4
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
dbfread==2.0.7
dotenv==0.9.9
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.3.0
PyMySQL==1.1.1
PyPDF2==3.0.1
python-dotenv==1.1.1
reportlab==4.4.3
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
